+++ narrenschiff/secretmap.py
@@ -3,6 +3,7 @@ import re
 import sys
 import uuid
 import shutil
+import difflib
 import subprocess
 from contextlib import suppress
 
@@ -92,14 +93,8 @@ class Secretmap(metaclass=Singleton):
         :return: Void
         :rtype: ``None``
         """
-        src = self._get_treasure_path(treasure)
-
-        with open(src, 'r') as f:
-            cipher = AES256Cipher(self.keychain)
-            enc_file_core = cipher.decrypt(f.read())
-
         with open(dest, 'w') as f:
-            f.write(enc_file_core)
+            f.write(self._decrypt(treasure))
 
     def peek(self, treasure):
         """
@@ -110,11 +105,7 @@ class Secretmap(metaclass=Singleton):
         :return: Void
         :rtype: ``None``
         """
-        src = self._get_treasure_path(treasure)
-
-        with open(src, 'r') as f:
-            cipher = AES256Cipher(self.keychain)
-            print(cipher.decrypt(f.read()))
+        click.echo(self._decrypt(treasure))
 
     def find(self, match, treasure):
         """
@@ -127,12 +118,7 @@ class Secretmap(metaclass=Singleton):
         :return: Void
         :rtype: ``None``
         """
-        src = self._get_treasure_path(treasure)
-
-        with open(src, 'r') as f:
-            logger.debug(f'Decrypting secretmap on {src}')
-            cipher = AES256Cipher(self.keychain)
-            secretmap = cipher.decrypt(f.read()).split("\n")
+        secretmap = self._decrypt(treasure).split('\n')
 
         logger.debug(f'Searching for "{match}"')
         for index, line in enumerate(secretmap, start=1):
@@ -142,6 +128,30 @@ class Secretmap(metaclass=Singleton):
                 prefix = (f'\033[35m{treasure}\033[0m:\033[32m{index}\033[0m')
                 formatted = line.replace(result, f'\033[31m{result}\033[0m')
                 print(f'{prefix}:{formatted}')
+
+    def diff(self, secretmaps):
+        """
+        Compare secretmaps line by line.
+
+        :param secretmaps: Two secretmaps that should be compared
+        :type secretmaps: ``tuple``
+        :return: Void
+        :rtype: ``None``
+        """
+        differences = difflib.unified_diff(
+            self._decrypt(secretmaps[0]).splitlines(keepends=True),
+            self._decrypt(secretmaps[1]).splitlines(keepends=True),
+            fromfile=secretmaps[0],
+            tofile=secretmaps[1]
+        )
+
+        for difference in differences:
+            if difference.startswith('-'):
+                click.secho(difference, nl=False, fg='red')
+            elif difference.startswith('+'):
+                click.secho(difference, nl=False, fg='green')
+            else:
+                click.secho(difference, nl=False)
 
     def destroy(self, treasure):
         """
@@ -221,8 +231,8 @@ class Secretmap(metaclass=Singleton):
         """
         Get path to treasure from config file.
 
-        :param config: Name of the treasure
-        :type config: ``str``
+        :param treasure: Name of the treasure
+        :type treasure: ``str``
         :return: Path to encrypted secretmap file
         :rtype: ``str``
         """
@@ -239,6 +249,12 @@ class Secretmap(metaclass=Singleton):
             sys.exit(1)
 
     def _read_config(self):
+        """
+        Load config file from the filesystem.
+
+        :return: Content of the config file
+        :rtype: ``dict``
+        """
         try:
             with open(self.filepath, 'r') as f:
                 config = yaml.load(f, Loader=yaml.FullLoader)
@@ -248,5 +264,31 @@ class Secretmap(metaclass=Singleton):
         return config if config else {}
 
     def _write_config(self, config):
+        """
+        Write config file to filesystem as YAML file.
+
+        :param config: Content of the config file
+        :type config: ``dict``
+        :return: Void
+        :rtype: ``None``
+        """
         with open(self.filepath, 'w') as f:
             f.write(yaml.dump(config))
+
+    def _decrypt(self, treasure):
+        """
+        Decrypt treasure, and return it as a cleartext string.
+
+        :param treasure: Name of the treasure
+        :type treasure: ``str``
+        :return: Cleartext string
+        :rtype: ``str``
+        """
+        src = self._get_treasure_path(treasure)
+
+        with open(src, 'r') as f:
+            logger.debug(f'Decrypting secretmap on {src}')
+            cipher = AES256Cipher(self.keychain)
+            secretmap = cipher.decrypt(f.read())
+
+        return secretmap